+++ lib/shark/client/connection.rb
@@ -1,14 +1,17 @@
 module Shark
   module Client
     class Connection
+      HTTP_METHODS_WITH_BODY = [:post, :patch, :put]
+
       attr_reader :site, :connection_options
 
       def initialize(options = {})
         @site = options.fetch(:site)
-        @connection_options = options.slice(:app_name, :params)
+        @connection_options = {}
         @connection_options[:headers] = options.fetch(:headers)
 
         @connection = Faraday.new do |faraday|
+          faraday.use Shark::Middleware::ComposeRequest
           faraday.use Shark::Middleware::Status
           faraday.use JsonApiClient::Middleware::ParseJson
           faraday.adapter :net_http_persistent
@@ -18,27 +21,44 @@ module Shark
       def use(middleware, *args, &block); end
 
       def run(action, path, params = {}, headers = {})
-        raise ArgumentError, 'Configuration :site cannot be nil' if site.blank?
-        raise ArgumentError, 'Parameter :path cannot be nil' if path.blank?
+        raise ArgumentError, "Configuration :site cannot be nil"  if site.blank?
+        raise ArgumentError, "Parameter :path cannot be nil"      if path.blank?
 
         url = File.join(site, path)
-        options = request_options(params, headers)
+        headers = request_headers(headers)
+        params = request_params(params)
 
-        BimaHttp.request(action, url, options)
+        @connection.send(action) do |request|
+          request.url(url)
+          request.headers.merge!(headers)
+          if request_with_body?(action)
+            request.body = params
+          else
+            request.params.merge!(params)
+          end
+        end
       end
+
 
       private
 
-      def request_options(params = {}, headers = {})
-        options = connection_options.reverse_merge(params: {}, headers: {})
-        options[:headers] = options[:headers].merge(headers)
-        if Shark._service_token.present?
-          options[:headers].merge!('Authorization' => "Bearer #{Shark._service_token}")
+      def request_headers(req_headers = {})
+        headers = connection_options[:headers] || {}
+        headers = headers.merge(req_headers)
+        if Shark.service_token.present?
+          headers["Authorization"] = "Bearer #{Shark.service_token}"
         end
-        options[:params] = options[:params].merge(params)
-        options[:connection] = @connection
 
-        options
+        headers
+      end
+
+      def request_params(req_params = {})
+        params = connection_options[:params] || {}
+        params.merge(req_params)
+      end
+
+      def request_with_body?(action)
+        HTTP_METHODS_WITH_BODY.include?(action.to_sym)
       end
     end
   end