+++ lib/shark/concerns/normalized_email.rb
@@ -1,0 +1,12 @@
+module Shark
+  module NormalizedEmail
+    extend ActiveSupport::Concern
+
+    class_methods do
+      def normalize_email(email)
+        email.to_s.downcase.strip
+      end
+    end
+  end
+end
+