+++ lib/shark/contact_service/contact.rb
@@ -1,7 +1,7 @@
 module Shark
   module ContactService
     class Contact < Base
-      extend Shark::ContactService::Concerns::NormalizedEmail
+      include Shark::NormalizedEmail
 
       custom_endpoint :avatar, on: :collection, request_method: :get
 
@@ -13,7 +13,7 @@ module Shark
       def self.find_by_email(email)
         normalized_email = normalize_email(email)
         options = { filter: { equals: { email: normalized_email } } }
-        where(options)
+        where(options).first
       end
 
       def account