+++ lib/shark/form_service/v2/base.rb
@@ -1,0 +1,12 @@
+module Shark
+  module FormService
+    module V2
+      class Base < FormService::Base
+        def self.site
+          File.join(::Shark.configuration.form_service.site, "/api/v2")
+        end
+      end
+    end
+  end
+end
+