+++ lib/shark/form_service/v2/form_input.rb
@@ -1,0 +1,14 @@
+module Shark
+  module FormService
+    module V2
+      class FormInput < V2::Base
+        belongs_to :form
+
+        def self.table_name
+          "inputs"
+        end
+      end
+    end
+  end
+end
+