+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 ## Changelog
 
+#### 0.6.2
+- [break] remove `URI.join()` from `settings.rb` in apps
+- contact service:
+    - supports activity
+- form service:
+    - supports API version 2
+
 #### 0.6.1
 - fixes including of RSpec helper files